add editorconfig file, remove query from UserRepository, create packages, make variables final in CustomUserDetails and change annotation in CustomUserDetailsService
