--- src/main/java/com/gym/GymApp/AppController.java
@@ -1,45 +1,0 @@
-package com.gym.GymApp;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-
-import java.util.List;
-
-@Controller
-public class AppController {
-
-    @Autowired
-    private UserRepository repository;
-
-    @GetMapping("")
-    public String viewHomePage(){
-        return "index";
-    }
-
-    @GetMapping("/register")
-    public String showSignUpForm(Model model){
-        model.addAttribute("user", new User());
-
-        return "signup_form";
-    }
-
-    @PostMapping("/process_register")
-    public String processingRegistration(User user){
-        BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
-        String encodedPassword = bCryptPasswordEncoder.encode(user.getPassword());
-        user.setPassword(encodedPassword);
-        repository.save(user);
-        return "register_success";
-    }
-
-    @GetMapping("/list_users")
-    public String viewUsersList(Model model){
-        List<User> userList = repository.findAll();
-        model.addAttribute("userList",userList);
-        return "users";
-    }
-}
