+++ src/main/java/com/gym/GymApp/auth/WebSecurityConfig.java
@@ -1,0 +1,64 @@
+package com.gym.GymApp.auth;
+
+import com.gym.GymApp.repository.UserRepository;
+import com.gym.GymApp.service.CustomUserDetailsService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+import javax.sql.DataSource;
+
+@Configuration
+@EnableWebSecurity
+public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+
+	@Autowired
+	private DataSource dataSource;
+	@Autowired
+	private UserRepository repository;
+
+	@Bean
+	public UserDetailsService userDetailsService() {
+		return new CustomUserDetailsService(repository);
+	}
+
+	@Bean
+	public BCryptPasswordEncoder bCryptPasswordEncoder() {
+		return new BCryptPasswordEncoder();
+	}
+
+	@Bean
+	public DaoAuthenticationProvider authenticationProvider() {
+		DaoAuthenticationProvider authenticationProvider = new DaoAuthenticationProvider();
+		authenticationProvider.setUserDetailsService(userDetailsService());
+		authenticationProvider.setPasswordEncoder(bCryptPasswordEncoder());
+
+		return authenticationProvider;
+	}
+
+	@Override
+	protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+		auth.authenticationProvider(authenticationProvider());
+	}
+
+	@Override
+	protected void configure(HttpSecurity http) throws Exception {
+		http.authorizeRequests()
+			.antMatchers("/list_users").authenticated()
+			.anyRequest().permitAll()
+			.and()
+			.formLogin()
+			.usernameParameter("email")
+			.defaultSuccessUrl("/list_users")
+			.permitAll()
+			.and()
+			.logout().logoutSuccessUrl("/").permitAll();
+	}
+}
