+++ src/main/java/com/gym/GymApp/repository/UserRepository.java
@@ -1,10 +1,10 @@
-package com.gym.GymApp;
+package com.gym.GymApp.repository;
 
+import com.gym.GymApp.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    @Query("SELECT u FROM User u WHERE u.email = ?1")
-    User findByEmail(String email);
+	User findByEmail(String email);
 }
